``` chore(setup): update version to 0.1.8 and enhance command line interface
A new verbose flag (-V, --verbose) has been added, allowing users to enable verbose output for more detailed information during execution. This is particularly useful for debugging and understanding the flow of the program.

Furthermore, the version flag (-v, --version) has been updated to display the new version number when invoked. This ensures that users can easily check which version of 'super-grep' they are currently using. Currently hard-coded at top of file

Additionally, a debug print statement has been added to indicate when all items in the search queue are being displayed, enhancing the debugging experience. The handling of the depth argument has also been refined to ensure that a value of -1 is correctly interpreted as None, preventing potential issues during execution for infinite depth searches.

Fixes: Cannot find anything with version 0.15 (Issue #1)
This commit fixes the actual issue the user was experiencing, where --depth parameter was not handled corrected for the case of "-1" infinite depth
#1